NES: Merge bank switching changes to sdcc_4.3

+++ gbdk-lib/examples/cross-platform/banks_farptr/src/bank2code.c
@@ -15,7 +15,7 @@ void some_bank2_proc0(void) __banked {
 }
 
 BANKREF(some_bank2_proc1)
-int some_bank2_proc1(int param1, int param2) __banked {
+int some_bank2_proc1(uint8_t param1, uint8_t param2) __banked {
     printf("some_bank2_proc1\n");
     return local_bank2_proc(param1, param2);
 }
